requirements.txt
fastapi
uvicorn[standard]
pyannote.audio
openai-whisper
pydub
ffmpeg-python
torch
numpy
python-multipart